Toggle des headlines après le chargement des images
