+++ assets/_articles.scss
@@ -87,8 +87,5 @@
 }
 
 .collapsed {
-  visibility: hidden;
-  position: absolute !important;
-  top: 0;
-  z-index: -10;
+  @extend .dn;
 }
