+++ resources/_gen/assets/scss/en/main.scss_f300667da4f5b5f84e1a9e0702b2fdde.content
@@ -1938,7 +1938,7 @@ img {
      -l  = large
 
 */
-.dn, .slides__nav, .in-sidebar, body > section > aside {
+.dn, .slides__nav, .collapsed, .in-sidebar, body > section > aside {
   display: none; }
 
 .di, .bibliography-entry h3, header[role="banner"] .main-navigation li,
@@ -7246,12 +7246,6 @@ body > section [role="main"] {
 
 .clickable {
   cursor: pointer; }
-
-.collapsed {
-  visibility: hidden;
-  position: absolute !important;
-  top: 0;
-  z-index: -10; }
 
 .mt-auto, .cards-list .card .card__footer {
   margin-top: auto; }
